feat: 라우터 적용 및 CardRegister, CardRegisterCompleted 페이지 생성
