+++ src/pages/CardRegisterCompleted.tsx
@@ -1,0 +1,5 @@
+function CardRegisterCompleted() {
+  return <h1>CardRegisterCompleted</h1>;
+}
+
+export default CardRegisterCompleted;
